[finance_report_ui] fix(home): consolidate attention into one /attention queue + split useDashboardData (#1116, #1119)
EPIC-022 AC22.16. The everyday-user Home still leaked internal accounting
pipeline as first-class destinations, competing with the Trust Meter →
/attention signal (Axiom B asks for one confidence-ranked queue):

- AC22.16.1 (#1116): onboarding "getting started" steps now point only at
  everyday surfaces — first step is Upload, no step links to the
  accounting-jargon /accounts route.
- AC22.16.2 (#1116): the analytics "Risk radar" card and the unmatched-alerts
  CTA route to the unified /attention queue instead of parallel Advanced
  reconciliation internals (/reconciliation, /reconciliation/unmatched, /review).
- AC22.16.3 (#1119): decompose the ~294-line useDashboardData god-hook into
  composable, independently-usable hooks (useDashboardSnapshot + useAssetTrend);
  the aggregate hook preserves its existing public contract. Behavior-preserving,
  no backend or schema change.

Tests reference the new AC IDs; full FE suite + coverage thresholds green.
